+++ src/lib.rs
@@ -9,7 +9,7 @@
     stable_features
 )]
 #![cfg_attr(test, feature(test))]
-#![feature(portable_simd)]
+#![feature(portable_simd, try_trait_v2)]
 // Stabilized (in at least one channel) but left here in case it helps backwards
 // compatibility. Can be removed in the future.
 #![feature(const_fn_floating_point_arithmetic, let_chains, lazy_cell, slice_as_chunks)]