+++ CHANGELOG.md
@@ -9,6 +9,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 ### Added
 
 - Added `nw_scalar_score` and `nw_scalar_alignment` for performing global sequence alignment
+- Added the method `process_results` to more robustly work with iterators of results
 - Added `get_aligned_query`, similar to `get_aligned_seqs` but returning only the query sequence
 - Added `to_eq_x` to `Alignment` and `AlignmentStates`, which converts `M` to either `=` or `X` in the alignment
 - Added `from_ciglets_unchecked` and `from_cigar_unchecked` to `AlignmentStates`